requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langgraph==0.2.70
python-dateutil
openai==1.61.1
eventregistry==9.1
ddgs>=4.0.0
streamlit>=1.28.0